requirements.txt
pandas
numpy
langchain
langchain-community
pypdf
sentence-transformers
chromadb

# in main
mistralai 
streamlit 
faiss-cpu
python-dotenv